lab/pyproject.toml
[tool.poetry]
name = "lab"
version = "0.1.0"
description = ""
authors = ["surister <surister98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.2.1"
pymongo = "^4.8.0"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
